Remove upper limits on output image size
This commit removes the upper limits of 8192x4096 pixels output image
by allocating all memory dynamically and changing the square arrays to
an array of pointers to arrays of pixels (the columns).
This also reduces the memory consumption for smaller images, for example
for 640x480 from 141MB to 14MB. Tested up to 10000 x 10000.