src/waveform.c : Add option to specify outer border width.
User input is multiplied by 2, because line-width for the border goes
beyond the dimensions of the resulting image (by half). In other words,
to have a border of 1pt, you have to set line-width to 2pt, but we hide
that detail from the user. Similarly, altho the default in "-h" states 1.0f,
the actual default value is 2.0f.

Signed-off-by: Erik de Castro Lopo <erikd@mega-nerd.com>