Make FFT window function selectable with --kaiser and --nuttall options
The Kaiser window is noisy on certain input data. e.g.:
    sndfile-generate-chirp -amp 1.0 -linear -from 22000 -to 26000 \
	96000 10 chirp96.wav
    sndfile-spectrogram chirp96.wav 800 600 chirp96.png
Other window functions (Nuttall, Hanning) do not have this problem.

This patch add a --nuttall option to select the Nuttall window function,
whose code is already included in the source, as an alternative to the
default one, the Kaiser window. There is a --kaiser option for symmetry.

An alternative would be --window=nuttall so we can beat Audacity's ten (!)
different window functions but --nuttall is easier to parse (and to type!)