Fix range of gray values
It used to be that the gray values ran evenly from 0 to 254, generating 255
extremely rarely, when the -logdB of a pixel value was exactly 0.0, which
provoked a "sanity check" error.

This change makes it use 255 for the top range of values, representing a
range of intensity values the same size as the ranges represented by
each of the other levels of gray.