Fix the interference pattern "florets" artifact in the output.
calc_magnitude was incorrectly decoding the "half-complex" output format
used by FFTW, summing the real and imaginary parts of different
frequency bands instead of those for the same frequency band.

This change rectifies that out-by-one error and also sets magnitudes[0]
which previously was always left black for some reason.
It also moves calc_magnitudes() from common.c into spectrogram.c as it is
FFTW-specific and this in the only place it is used.