Rewrite axis-tick positioner
The old tick-positioner used to go crazy when labelling 0-13000 or 15000
giving labels at 0 2167 4333 6500 8667 10833 13000.

This change implements the "1, 2 or 5" rule, whereby the least significant
digit that changes from label to label should change by a
multiple of 1, 2 or 5.  The old code did this from 0 to 12000 Hz, and the
new code gives the same number of divisions as the old in this range, then
something sensible.

It also reacts well to ranges where the minimum is greater than 0, whereas
the old code only worked with ranges starting at 0.

The number-printing code also changes. It used to give:
- 0 decimal places from 10 upwards
- 1 decimal place from 1.0 to 9.99
- 2 decimal places below 1.0
It now prints the number of decimal places necessary to distinguish
between adjacent values, which we know from the choice of step value.
This is essential for ranges like 1000 to 1001 and avoids printing
"0 0.50 1.0 1.5 2.0".