Implement --log-freq option for log frequency axis
This puts a logarithmic axis on the frequency axis from --min-freq
(default 20Hz) to --max-freq (default the Nyquist frequency).

The current numbering/ticking algorithm for log axis works one way for
ranges of a factor of ten or more, numbering in one of:
1 - 10 - 100 - 1000 with ticks at 2 and 5 positions between them or
1 - 2 - 5 - 10 - 20 - 50 - 100 with ticks at 2-9, 20-90 and so on
For ranges less than a decade (max/min < 10), it uses the same numbering/
ticking scheme as for the linear axes, but spacing the ticks logarithmically.

Signed-off-by: Erik de Castro Lopo <erikd@mega-nerd.com>
Closes: https://github.com/erikd/sndfile-tools/issues/35
Closes: https://github.com/erikd/sndfile-tools/issues/36