Add --fft-freq flag to specify window length
This new flag allows you to adjust the timr/frequency precision tradeoff
by setting the FFT window size via a new flag --fft-freq=X which gives you
a window an Xth of a second wide, determining the lowest resolvable frequency
and the distance between the frequency bands in the linear spectrogram.

The default value is the old formula, whose comments say it resolves down to 20Hz
but, for a 44100Hz file gives an equivalent fft_freq of 8.963 Hz.

Signed-off-by: Erik de Castro Lopo <erikd@mega-nerd.com>