Fix log frequency axis labelling sometimes going crazy
The case producing the most ticks, 10-99999 or 11-100000, used to
overflow the maximum of 20 ticks. This ups the maximum to 40 giving,
for the usual case (CD audio) labels at 10,20,50,100,200,500... and
ticks at 30,40,60,70,80. An alternative, of reducing the number of
ticks, leaves 100-1000-10000 as labels and  ticks at 2 and 5 points
between them, which seems less clear.

Fixes github.com/erikd/issues/43