rewrite sndfile-mix-to-mono(1) on mdoc(7)
Currently, the manpages of sndfile-tools
are written using the legacy man(7) markup language.
I propose to rewrite them into the markup of mdoc(7),
(like with libsndfile some years ago, and
sndfile-resample which has since been moved here).

As an example, here is a rewrite of sndfile-mix-to-mono,
being the shortest of course :-)

The main advantage is that the mdoc(7) language,
available for decades alongside man(7), is semantical
and describes the intent, as opposed the the low-level
formatting of roff(7). For example:

.Fl w Op Ar port

means there is a -w option which takes an optional 'port' argument;
as opposed to

.BR \-w [ =\fIport ]

which says switch to bold, type a hyphen and a 'w',
then switch to italics, and type out 'port' in brackets.

Both languages are well supported since the eighties;
nowadays, the prevalent formatters are groff on the linuxes
and mandoc on the BSDs and macOS.